change leaderboard to use message embed
